Moving identity capabilities to a separate file. fixing some bugs
